remove bz2 from re file
